// api/generated/model/reportRequestDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportRequestDtoStatus } from './reportRequestDtoStatus';
import type { ReportTemplateDto } from './reportTemplateDto';

export interface ReportRequestDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  trainerId?: string | null;
  /** @nullable */
  traineeId?: string | null;
  /** @nullable */
  templateId?: string | null;
  status?: ReportRequestDtoStatus;
  /** @nullable */
  dueAt?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt?: string;
  /** @nullable */
  submittedAt?: string | null;
  template?: ReportTemplateDto;
}