// api/generated/model/reportTemplateFieldDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportTemplateFieldDtoType } from './reportTemplateFieldDtoType';

export interface ReportTemplateFieldDto {
  /** @nullable */
  key?: string | null;
  /** @nullable */
  label?: string | null;
  type?: ReportTemplateFieldDtoType;
  isRequired?: boolean;
  order?: number;
  moduleConfig?: unknown | null;
}